requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
kafka-python==2.0.2
apachelogs==0.13.0
user-agents==2.2.0
geoip2==4.8.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
boto3==1.34.144
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
prometheus-client==0.20.0
requests==2.32.3
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.11.2
apache-airflow-providers-amazon==8.23.0
elasticsearch==8.14.0
